Refactor pkg utils
